Fixed serialization/deserialization of bool type
The library couldn't serialize and deserialize boolean type as
bool because of one wrong condition in codec.py
(isinstance(bool(), int) return True). In that patch the condition
has been modified to resolve that issue.

Also the patch contains a couple of tests which check if
boolean serialization/deserialization work correct.